refactor: use multiple files for types
